♻️ Move simlayer to its own file
